requirements.txt
# 패키지 설치
# pip install -r requirements.txt
kafka-python
matplotlib
pandas
pyarrow
pymongo
pyspark
scikit-learn
seaborn